Update first_streamlit_app.py
ask our app to put the list of selected fruits into a variable called fruits_selected. Then, we'll ask our app to use the fruits in our fruits_selected list to pull rows from the full data set (and assign that data to a variable called fruits_to_show). Finally, we'll ask the app to use the data in fruits_to_show in the dataframe it displays on the page. 
1. Wrap the multiselect into "fruits_selected"
2. Create: fruits_to_show = my_fruit_list.loc[fruits_selected]
3. Call the dataframe fruits_to_show // streamlit.dataframe(fruits_to_show)